requirements.txt
numpy==1.23.4
opencv-python==4.5.5.64
Pillow==9.2.0
torch==1.11.0
torchvision==0.12.0
ninja
ftfy==6.1.1
regex==2022.9.13
tqdm==4.64.1
git+https://github.com/openai/CLIP.git
